# api-gateway/docs/swagger.yaml
...
openapi: 3.0.3
info:
  title: Granula API
  version: 1.0.0
  description: |
    # Granula API

    API для интеллектуального сервиса планирования ремонта и перепланировки квартир.

    ## Возможности

    - 🏠 **Загрузка планировок** - Загружайте изображения планировок для автоматического распознавания
    - 🤖 **AI распознавание** - Автоматическое распознавание комнат, стен и элементов
    - 🎨 **3D редактор** - Редактирование планировки в интерактивном 3D редакторе
    - ✅ **Проверка норм** - Автоматическая проверка на соответствие СНиП и ЖК РФ
    - 💡 **AI генерация** - Генерация вариантов планировки с помощью AI
    - 📋 **Заявки экспертам** - Подача заявок на консультации специалистов БТИ

    ## Аутентификация

    API использует JWT (JSON Web Tokens) для аутентификации.

    ### Получение токенов

    1. Зарегистрируйтесь через `POST /api/v1/auth/register`
    2. Или войдите через `POST /api/v1/auth/login`
    3. Получите `access_token` и `refresh_token`

    ### Использование токенов

    Передавайте Access Token в заголовке Authorization:
    ```
    Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    ```

    ### Обновление токенов

    Access Token действует 15 минут. Для обновления используйте `POST /api/v1/auth/refresh`

    ## Rate Limiting

    | Тип запроса | Лимит |
    |-------------|-------|
    | Анонимные | 60 req/min |
    | Аутентифицированные | 300 req/min |
    | AI endpoints | 30 req/min |

    При превышении лимита возвращается HTTP 429.

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен аутентификации. Получите через `/auth/login` или `/auth/register`.

        Формат: `Bearer <token>`

  # ===========================================================================
  # Schemas
  # ===========================================================================
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Машиночитаемый код ошибки
          example: VALIDATION_ERROR
        message:
          type: string
          description: Человекочитаемое описание ошибки
          example: Неверный формат email
        details:
          type: object
          additionalProperties:
            type: string
          description: Детали ошибок валидации по полям
          example:
            email: "Неверный формат email"
            password: "Минимум 8 символов"
        request_id:
          type: string
          description: ID запроса для отладки
          example: req-abc123def456
...
    ComplianceCheckRequest:
      type: object
      required:
        - scene_id
      properties:
        scene_id:
          type: string
          format: uuid
          description: ID сцены для проверки
        branch_id:
          type: string
          format: uuid
          description: ID ветки (опционально)
        categories:
          type: array
          items:
            type: string
          description: Категории правил для проверки
          example: [minimum_area, ventilation]
        operation_type:
          type: string
          enum: [construction, renovation, redevelopment]
          example: redevelopment
    ComplianceCheckResponse:
      type: object
      properties:
        check_id:
          type: string
          format: uuid
        scene_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [passed, warning, failed]
          description: |
            Общий статус:
            - **passed** - Все проверки пройдены
            - **warning** - Есть предупреждения
            - **failed** - Есть критические нарушения
          example: warning
        score:
          type: integer
          description: Общий балл соответствия (0-100)
          example: 78
        violation_count:
          type: integer
          example: 2
        warning_count:
          type: integer
          example: 3
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
        category_stats:
          type: array
          items:
            $ref: '#/components/schemas/CategoryStat'
        checked_at:
          type: string
          format: date-time
        processing_time_ms:
          type: integer
          example: 320
    Violation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rule_id:
          type: string
          example: SNIP-MIN-AREA-01
        rule_name:
          type: string
          example: Минимальная площадь жилой комнаты
        category:
          type: string
          example: minimum_area
        severity:
          type: string
          enum: [error, warning, info]
          description: |
            Серьезность:
            - **error** - Критическое нарушение, блокирует согласование
            - **warning** - Предупреждение, рекомендуется исправить
            - **info** - Информационное сообщение
          example: error
        description:
          type: string
          example: Площадь спальни (7.5 кв.м) меньше минимально допустимой (8 кв.м)
        affected_elements:
          type: array
          items:
            type: object
            properties:
              element_id:
                type: string
                format: uuid
              element_type:
                type: string
              element_name:
                type: string
        recommendations:
          type: array
          items:
            type: string
          example:
            - Увеличить площадь комнаты минимум на 0.5 кв.м
            - Рассмотреть объединение с соседней комнатой
        legal_reference:
          type: string
          example: СНиП 31-01-2003, п. 5.3
    CategoryStat:
      type: object
      properties:
        category:
          type: string
          example: minimum_area
        category_name:
          type: string
          example: Минимальные площади
        total_rules:
          type: integer
          example: 8
        passed:
          type: integer
          example: 6
        violations:
          type: integer
          example: 1
        warnings:
          type: integer
          example: 1
    ComplianceRule:
      type: object
      properties:
        id:
          type: string
          example: SNIP-MIN-AREA-01
        name:
          type: string
          example: Минимальная площадь жилой комнаты
        category:
          type: string
          example: minimum_area
        description:
          type: string
          example: Площадь жилой комнаты должна быть не менее 8 кв.м
        default_severity:
          type: string
          enum: [error, warning, info]
          example: error
        legal_reference:
          type: string
          example: СНиП 31-01-2003
        applicable_operations:
          type: array
          items:
            type: string
          example: [construction, redevelopment]
        is_active:
          type: boolean
          example: true

    # -------------------------------------------------------------------------
    # Request (Expert) Schemas
    # -------------------------------------------------------------------------
...
    ChatMessageRequest:
      type: object
      required:
        - message
      properties:
        session_id:
          type: string
          format: uuid
          description: ID сессии (создается автоматически)
        message:
          type: string
          minLength: 1
          maxLength: 4000
          example: Как лучше организовать пространство в квартире 50 кв.м?
        scene_id:
          type: string
          format: uuid
          description: ID сцены для контекста
    ChatMessageResponse:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        message_id:
          type: string
          format: uuid
        response:
          type: string
          example: Для квартиры 50 кв.м рекомендую...
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/AISuggestion'
        tokens_used:
          type: integer
          example: 450
        created_at:
          type: string
          format: date-time
    AISuggestion:
      type: object
      properties:
        type:
          type: string
          enum: [layout_change, furniture_placement, space_optimization, compliance_warning]
        title:
          type: string
          example: Объединение кухни и гостиной
        description:
          type: string
        confidence:
          type: number
          format: float
          example: 0.87
    GenerateVariantsRequest:
      type: object
      required:
        - scene_id
        - count
      properties:
        scene_id:
          type: string
          format: uuid
        count:
          type: integer
          minimum: 1
          maximum: 5
          example: 3
        style:
          type: string
          enum: [modern, classic, minimalist, scandinavian, loft]
          example: modern
        priorities:
          type: array
          items:
            type: string
          example: [maximize_light, open_space]
        keep_elements:
          type: array
          items:
            type: string
            format: uuid
          description: ID элементов, которые нельзя изменять
        budget_range:
          type: object
          properties:
            min:
              type: integer
              example: 100000
            max:
              type: integer
              example: 500000
    GenerateVariantsResponse:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [processing, completed, failed]
        variants:
          type: array
          items:
            $ref: '#/components/schemas/GeneratedVariant'
        processing_time_ms:
          type: integer
        error:
          type: string
    GeneratedVariant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Вариант 1 - Открытое пространство
        description:
          type: string
        branch_id:
          type: string
          format: uuid
        quality_score:
          type: integer
          description: Оценка качества (0-100)
          example: 85
        estimated_cost:
          type: integer
          description: Оценочная стоимость реализации
          example: 350000
        preview_url:
          type: string
          format: uri
        changes:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              description:
                type: string
              affected_elements:
                type: array
                items:
                  type: string

    # -------------------------------------------------------------------------
    # Common Schemas
    # -------------------------------------------------------------------------
...
paths:
...
  /compliance/check:
    post:
      tags: [compliance]
      summary: Проверить соответствие нормам
      operationId: checkCompliance
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceCheckRequest'
      responses:
        '200':
          description: Результат проверки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceCheckResponse'
  /compliance/rules:
    get:
      tags: [compliance]
      summary: Список правил
      operationId: listComplianceRules
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: operation_type
          in: query
          schema:
            type: string
            enum: [construction, renovation, redevelopment]
      responses:
        '200':
          description: Список правил
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: '#/components/schemas/ComplianceRule'
                  total_rules:
                    type: integer

  # ---------------------------------------------------------------------------
  # AI Endpoints
  # ---------------------------------------------------------------------------
  /ai/chat:
    post:
      tags: [ai]
      summary: Отправить сообщение AI
      operationId: sendChatMessage
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
      responses:
        '200':
          description: Ответ AI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
        '429':
          description: Превышен лимит запросов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ai/generate:
    post:
      tags: [ai]
      summary: Сгенерировать варианты
      operationId: generateVariants
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateVariantsRequest'
      responses:
        '202':
          description: Генерация запущена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateVariantsResponse'

  # ---------------------------------------------------------------------------
  # Request Endpoints
  # ---------------------------------------------------------------------------
...