# api-gateway/internal/handlers/ai_handler.go
// =============================================================================
// AI Handler - HTTP handlers for AI operations.
// =============================================================================
package handlers

import (
	"encoding/json"
	"net/http"

	"github.com/xiiisorate/granula_api/api-gateway/internal/dto"
)

// =============================================================================
// AIHandler handles AI-related HTTP requests.
// =============================================================================

// AIHandler provides HTTP handlers for AI operations.
type AIHandler struct {
	// aiClient pb.AIServiceClient
}

// NewAIHandler creates a new AIHandler.
func NewAIHandler() *AIHandler {
	return &AIHandler{}
}

// =============================================================================
// Chat Endpoints
// =============================================================================

// SendChatMessage godoc
// @Summary Отправить сообщение в AI чат
// @Description Отправляет сообщение в AI чат и получает ответ
// @Tags ai
// @Accept json
// @Produce json
// @Param request body dto.ChatMessageRequest true "Сообщение"
// @Success 200 {object} dto.ChatMessageResponse "Ответ AI"
// @Failure 400 {object} dto.ErrorResponse "Ошибка валидации"
// @Failure 401 {object} dto.ErrorResponse "Не авторизован"
// @Failure 429 {object} dto.ErrorResponse "Превышен лимит запросов"
// @Failure 500 {object} dto.ErrorResponse "Ошибка AI сервиса"
// @Security BearerAuth
// @Router /ai/chat [post]
func (h *AIHandler) SendChatMessage(w http.ResponseWriter, r *http.Request) {
	var req dto.ChatMessageRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondError(w, http.StatusBadRequest, "INVALID_JSON", "Invalid request body")
		return
	}

	// TODO: Call AI gRPC service

	response := dto.ChatMessageResponse{
		SessionID: "session-placeholder-id",
		MessageID: "message-placeholder-id",
		Response:  "Для квартиры 50 кв.м рекомендую рассмотреть следующие варианты планировки...",
		TokensUsed: 150,
	}

	respondJSON(w, http.StatusOK, response)
}

// GetChatHistory godoc
// @Summary История чата
// @Description Возвращает историю сообщений в сессии чата
// @Tags ai
// @Accept json
// @Produce json
// @Param session_id path string true "ID сессии" format(uuid)
// @Param limit query int false "Количество сообщений" default(50) maximum(100)
// @Param before query string false "ID сообщения для пагинации назад"
// @Success 200 {object} dto.ChatHistoryResponse "История чата"
// @Failure 401 {object} dto.ErrorResponse "Не авторизован"
// @Failure 404 {object} dto.ErrorResponse "Сессия не найдена"
// @Security BearerAuth
// @Router /ai/chat/{session_id}/history [get]
func (h *AIHandler) GetChatHistory(w http.ResponseWriter, r *http.Request) {
	// TODO: Call AI gRPC service

	response := dto.ChatHistoryResponse{
		SessionID:     "session-id",
		Messages:      []dto.ChatHistoryMessageResponse{},
		TotalMessages: 0,
	}

	respondJSON(w, http.StatusOK, response)
}

// DeleteChatSession godoc
// @Summary Удалить сессию чата
// @Description Удаляет сессию чата и всю историю сообщений
// @Tags ai
// @Accept json
// @Produce json
// @Param session_id path string true "ID сессии" format(uuid)
// @Success 200 {object} dto.SuccessResponse "Сессия удалена"
// @Failure 401 {object} dto.ErrorResponse "Не авторизован"
// @Failure 404 {object} dto.ErrorResponse "Сессия не найдена"
// @Security BearerAuth
// @Router /ai/chat/{session_id} [delete]
func (h *AIHandler) DeleteChatSession(w http.ResponseWriter, r *http.Request) {
	// TODO: Call AI gRPC service

	respondJSON(w, http.StatusOK, dto.SuccessResponse{
		Success: true,
		Message: "Chat session deleted",
	})
}

// =============================================================================
// Generation Endpoints
// =============================================================================

// GenerateVariants godoc
// @Summary Сгенерировать варианты планировки
// @Description Запускает AI генерацию вариантов планировки на основе существующей сцены
// @Tags ai
// @Accept json
// @Produce json
// @Param request body dto.GenerateVariantsRequest true "Параметры генерации"
// @Success 202 {object} dto.GenerateVariantsResponse "Задача запущена"
// @Failure 400 {object} dto.ErrorResponse "Ошибка валидации"
// @Failure 401 {object} dto.ErrorResponse "Не авторизован"
// @Failure 404 {object} dto.ErrorResponse "Сцена не найдена"
// @Failure 429 {object} dto.ErrorResponse "Превышен лимит запросов"
// @Security BearerAuth
// @Router /ai/generate [post]
func (h *AIHandler) GenerateVariants(w http.ResponseWriter, r *http.Request) {
	var req dto.GenerateVariantsRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondError(w, http.StatusBadRequest, "INVALID_JSON", "Invalid request body")
		return
	}

	// TODO: Call AI gRPC service

	response := dto.GenerateVariantsResponse{
		TaskID: "task-placeholder-id",
		Status: "processing",
	}

	respondJSON(w, http.StatusAccepted, response)
}

// GetGenerationStatus godoc
// @Summary Статус генерации
// @Description Возвращает статус задачи генерации вариантов
// @Tags ai
// @Accept json
// @Produce json
// @Param task_id path string true "ID задачи" format(uuid)
// @Success 200 {object} dto.GenerateVariantsResponse "Статус задачи"
// @Failure 401 {object} dto.ErrorResponse "Не авторизован"
// @Failure 404 {object} dto.ErrorResponse "Задача не найдена"
// @Security BearerAuth
// @Router /ai/generate/{task_id} [get]
func (h *AIHandler) GetGenerationStatus(w http.ResponseWriter, r *http.Request) {
	// TODO: Call AI gRPC service

	response := dto.GenerateVariantsResponse{
		TaskID: "task-id",
		Status: "completed",
		Variants: []dto.GeneratedVariantResponse{
			{
				ID:           "variant-1",
				Name:         "Вариант 1 - Открытое пространство",
				Description:  "Объединение кухни и гостиной",
				BranchID:     "branch-1",
				QualityScore: 85,
			},
		},
		ProcessingTime: 5200,
	}

	respondJSON(w, http.StatusOK, response)
}

// =============================================================================
// Recognition Endpoints
// =============================================================================

// RecognizeFloorPlan godoc
// @Summary Распознать планировку
// @Description Запускает AI распознавание изображения планировки
// @Tags ai
// @Accept multipart/form-data
// @Produce json
// @Param file formData file true "Изображение планировки"
// @Param workspace_id formData string true "ID воркспейса"
// @Success 202 {object} dto.RecognitionStatusResponse "Задача запущена"
// @Failure 400 {object} dto.ErrorResponse "Ошибка валидации"
// @Failure 401 {object} dto.ErrorResponse "Не авторизован"
// @Failure 413 {object} dto.ErrorResponse "Файл слишком большой"
// @Failure 429 {object} dto.ErrorResponse "Превышен лимит запросов"
// @Security BearerAuth
// @Router /ai/recognize [post]
func (h *AIHandler) RecognizeFloorPlan(w http.ResponseWriter, r *http.Request) {
	// Parse multipart form
	if err := r.ParseMultipartForm(10 << 20); err != nil {
		respondError(w, http.StatusBadRequest, "INVALID_FORM", "Failed to parse form data")
		return
	}

	// TODO: Call AI gRPC service

	response := dto.RecognitionStatusResponse{
		TaskID: "task-placeholder-id",
		Status: "queued",
	}

	respondJSON(w, http.StatusAccepted, response)
}

// GetRecognitionResult godoc
// @Summary Результат распознавания
// @Description Возвращает результат распознавания планировки
// @Tags ai
// @Accept json
// @Produce json
// @Param task_id path string true "ID задачи" format(uuid)
// @Success 200 {object} dto.RecognitionStatusResponse "Статус и результат"
// @Failure 401 {object} dto.ErrorResponse "Не авторизован"
// @Failure 404 {object} dto.ErrorResponse "Задача не найдена"
// @Security BearerAuth
// @Router /ai/recognize/{task_id} [get]
func (h *AIHandler) GetRecognitionResult(w http.ResponseWriter, r *http.Request) {
	// TODO: Call AI gRPC service

	response := dto.RecognitionStatusResponse{
		TaskID:   "task-id",
		Status:   "completed",
		Progress: 100,
	}

	respondJSON(w, http.StatusOK, response)
}